change data Type, seed order  - FK Error

+++ database/migrations/2016_02_10_180748_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
             $table->string('avatar',128);
             $table->string('remember_token',128);
             $table->string('social_token',128);
-            $table->tinyInteger('osk_id')->unsigned()->index();
+            $table->integer('osk_id')->unsigned()->index();
             $table->string('is_admin', 1);
             $table->timestamps();
         });
