+++ database/migrations/2016_02_11_091717_create_hours_table.php
@@ -15,10 +15,10 @@ class CreateHoursTable extends Migration
         Schema::create('hours', function (Blueprint $table) {
             //$table->increments('id')->unsigned();
             $table->date('drive_date')->primary();
-            $table->tinyInteger('student_id')->unsigned()->index();
+            $table->integer('student_id')->unsigned()->index();
             $table->foreign('student_id')->references('id')->on('students');
             $table->decimal('count', 2, 1);
-            $table->tinyInteger('instructor_id')->unsigned()->index();
+            $table->integer('instructor_id')->unsigned()->index();
             $table->foreign('instructor_id')->references('id')->on('instructors');
             $table->timestamps();
         });
